glass: use 'local' endpoints instead of 'orch'
The backend moved a bunch of endpoints to '/api/local', away from
'/api/orch'. Map those changes.
