+++ src/glass/src/app/core/dashboard/widgets/volumes-dashboard-widget/volumes-dashboard-widget.component.spec.ts
@@ -7,7 +7,10 @@ import { of } from 'rxjs';
 
 import { DashboardModule } from '~/app/core/dashboard/dashboard.module';
 import { VolumesDashboardWidgetComponent } from '~/app/core/dashboard/widgets/volumes-dashboard-widget/volumes-dashboard-widget.component';
-import { OrchService, Volume } from '~/app/shared/services/api/orch.service';
+import {
+  LocalNodeService,
+  Volume
+} from '~/app/shared/services/api/local.service';
 
 describe('DevicesDashboardWidgetComponent', () => {
   let component: VolumesDashboardWidgetComponent;
@@ -202,7 +205,8 @@ describe('DevicesDashboardWidgetComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(VolumesDashboardWidgetComponent);
     component = fixture.componentInstance;
-    spyOn(TestBed.inject(OrchService), 'volumes').and.returnValue(of(mockedVolumes));
+    spyOn(TestBed.inject(LocalNodeService), 'volumes')
+    .and.returnValue(of(mockedVolumes));
     fixture.detectChanges();
   });
