+++ src/glass/src/app/shared/services/api/orch.service.spec.ts
@@ -43,9 +43,4 @@ describe('OrchService', () => {
     expect(req.request.method).toBe('GET');
   });
 
-  it('should call facts', () => {
-    service.facts().subscribe();
-    const req = httpTesting.expectOne('api/orch/facts');
-    expect(req.request.method).toBe('GET');
-  });
 });
